Refactor: moving `Verbatim` et al. to `lexing/utils`.
