+++ src/lib.rs
@@ -12,17 +12,19 @@
 //! user *where* the error happened.
 //!
 
-// This will have to be removed to solve #5
+// This will have to be removed to solve #5:
 #![allow(incomplete_features)]
 #![feature(adt_const_params)]
 
 pub mod common;
 pub mod lexing;
 
+pub(crate) use avjason_macros::*;
+
 mod macro_test {
     use std::marker::PhantomData;
 
-    use avjason_macros::{ECMARef, Spanned, SpecRef};
+    use super::{ECMARef, Spanned, SpecRef};
 
     use crate::common::Span;
